requirements.txt
# Core dependencies
python-dotenv>=1.0.0    # For environment variable management
openai>=1.6.0           # For Azure OpenAI API interaction
crewai>=0.28.0          # For agent and task management
sympy>=1.12             # For safe mathematical evaluation
requests>=2.31.0        # For HTTP requests in functions

# Optional dependencies that may be needed for various tools
aiohttp>=3.8.4          # For async HTTP requests
tiktoken>=0.5.1         # For token counting with OpenAI models
